backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
DAWG-Python==0.7.2
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.5.0
gunicorn==22.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
pip==24.0
psycopg2-binary==2.9.9
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
regex==2024.5.15
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.0
Werkzeug==3.0.3
